Simplify states, increase use of state obj
Simplify state conversions, use State objects instead of np.ndarray in most places
