+++ examples/shockbox/shockbox.py
@@ -24,21 +24,21 @@ settings = {
     "interface_interpolation": "arithmetic_average",
     "reconstruction_type": "conservative",
     "upwind_mode": "primitive",
-    "write_solution": True,
+    "write_solution": False,
     "write_solution_mode": "every_n_timesteps",
     "write_solution_name": "shockbox",
     "write_every_n_timesteps": 30,
     "plot_every": 10,
     "CFL": 0.4,
     "t_final": 2.0,
-    "realplot": False,
+    "realplot": True,
     "profile": False,
     "gamma": 1.4,
     "rho_inf": 1.0,
     "a_inf": 343.0,
     "R": 287.0,
-    "nx": 500,
-    "ny": 500,
+    "nx": 50,
+    "ny": 50,
     "nghost": 1,
     "use_JIT": True,
 }
