+++ pyHype/flux/base.py
@@ -20,7 +20,7 @@ os.environ["NUMPY_EXPERIMENTAL_ARRAY_FUNCTION"] = "0"
 import numpy as np
 import numba as nb
 from abc import abstractmethod
-from pyHype.states.states import PrimitiveState, RoePrimitiveState
+from pyHype.states.primitive import PrimitiveState, RoePrimitiveState
 
 
 class FluxFunction:
