+++ pyHype/states/__init__.py
@@ -14,4 +14,5 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 from .base import State
-from .states import PrimitiveState, ConservativeState, RoePrimitiveState
+from .primitive import PrimitiveState, RoePrimitiveState
+from .conservative import ConservativeState
